requirements.txt
python-dotenv
discord.py
gspread